mcp_servers/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database dependencies
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.0.3

# Security and authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==41.0.7

# HTTP client and utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# JSON handling
orjson==3.9.10
jsonschema==4.20.0

# Validation and utilities
email-validator==2.1.0
python-slugify==8.0.1

# Testing dependencies (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality tools (dev)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Development tools
pre-commit==3.6.0

# MCP specific dependencies
# Note: Add actual MCP SDK when available
jsonrpc-async==2.1.2
websockets==12.0

# Rate limiting and middleware
slowapi==0.1.9

# CORS and security middleware
# (FastAPI includes these, but listing for clarity)

# Configuration and environment
dynaconf==3.2.4

# Date and time utilities
python-dateutil==2.8.2

# Async task queue (optional)
celery==5.3.4
redis[hiredis]==5.0.1
aioredis==2.0.1

# File handling
python-magic==0.4.27

# UUID utilities
uuid7==0.1.0

# Process monitoring
psutil==5.9.6